backend/requirements.txt
asgiref==3.3.4
Babel==2.9.1
backcall==0.2.0
decorator==5.0.9
Django==3.2.3
django-cors-headers==3.7.0
django-money==2.0
djangorestframework==3.12.4
ipython==7.24.1
ipython-genutils==0.2.0
jedi==0.18.0
matplotlib-inline==0.1.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
psycopg2==2.9.1
ptyprocess==0.7.0
py-moneyed==1.2
Pygments==2.9.0
pytz==2021.1
sqlparse==0.4.1
traitlets==5.0.5
wcwidth==0.2.5
